# api-spec.yaml
openapi: 3.0.1
info:
  title: Chattrix API
  description: |
    API toàn diện cho ứng dụng Chattrix, bao gồm xác thực, quản lý người dùng, và tính năng gọi điện thời gian thực.

    ---

    ### Chi tiết các sự kiện WebSocket
    Tất cả các tin nhắn từ server đều được bọc trong một đối tượng `WebSocketMessage` chung.
    
    **Message Events:**
    *   **`chat.message`**: Tin nhắn trò chuyện mới. (Payload: `OutgoingMessageDto`)
    *   **`message.updated`**: Một tin nhắn đã được cập nhật. (Payload: `MessageUpdateDto`)
    *   **`message.deleted`**: Một tin nhắn đã bị xóa. (Payload: `MessageDeleteDto`)
    *   **`message.reaction`**: Reaction đã được thêm/xóa khỏi tin nhắn. (Payload: `ReactionEventDto`)
    
    **Conversation Events:**
    *   **`conversation.update`**: Thông tin cuộc trò chuyện đã được cập nhật (ví dụ: tin nhắn cuối cùng). (Payload: `ConversationUpdateDto`)
    *   **`typing.indicator`**: User đang typing trong conversation. (Payload: `TypingIndicatorResponseDto`)
    
    **User Events:**
    *   **`user.status`**: Trạng thái trực tuyến của người dùng đã thay đổi. (Payload: `UserStatusDto`)
    
    **Friend Request Events:**
    *   **`friend.request.received`**: Nhận được friend request mới. (Payload: `FriendRequestResponse`)
    *   **`friend.request.accepted`**: Friend request được chấp nhận. (Payload: `FriendRequestResponse`)
    *   **`friend.request.rejected`**: Friend request bị từ chối. (Payload: `FriendRequestRejectDto`)
...
    
...
    
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ChatMessageRequest:
      type: object
      required: [content]
      properties:
        content: { type: string }
        type: { type: string, enum: [TEXT, IMAGE, VIDEO, AUDIO, FILE, LOCATION], default: TEXT }
        replyToMessageId: { type: integer, format: int64, nullable: true }
        mentions:
          type: array
          items:
            type: integer
            format: int64
        mediaUrl: { type: string, format: uri, nullable: true }
        thumbnailUrl: { type: string, format: uri, nullable: true }
        fileName: { type: string, nullable: true }
        fileSize: { type: integer, format: int64, nullable: true }
        duration: { type: integer, nullable: true }
        latitude: { type: number, format: double, nullable: true }
        longitude: { type: number, format: double, nullable: true }
        locationName: { type: string, nullable: true }
    UpdateMessageRequest:
      type: object
      required: [content]
      properties:
        content: { type: string, description: "Nội dung mới của tin nhắn." }

    # --- Call Schemas ---
...
    ApiResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
        data: { nullable: true }
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: "Only present for validation errors (e.g., 'Validation failed'). Single business errors don't have this field."
          example: "Validation failed"
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: "Error code (e.g., VALIDATION_ERROR, UNAUTHORIZED, USER_NOT_FOUND)"
              example: "RESOURCE_NOT_FOUND"
            message:
              type: string
              description: "Human-readable error message"
              example: "The requested resource was not found"
            details:
              type: object
              description: "Map of field names to error messages (only for validation errors with multiple field errors)"
              additionalProperties:
                type: string
              example:
                email: "Email already exists"
                username: "Username must be 4-20 characters"
        requestId:
          type: string
          description: "Unique request ID for tracking"
          example: "req-12345-abcde"
...
    ApiResponse_MessageResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
        data: { $ref: '#/components/schemas/MessageResponse' }
    ApiResponse_MessageListResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
        data:
          type: array
          items:
            $ref: '#/components/schemas/MessageResponse'
...
    ReadReceiptResponse:
      type: object
      properties:
        userId: { type: integer, format: int64 }
        username: { type: string }
        fullName: { type: string }
        avatarUrl: { type: string, format: uri, nullable: true }
        readAt: { type: string, format: date-time }

    # --- Reaction Schemas ---
...
    ApiResponse_ReadReceiptListResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReadReceiptResponse'
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/conversations/{conversationId}/messages:
    get:
      summary: "Get conversation messages"
      description: "Lấy danh sách tin nhắn trong một cuộc trò chuyện, có hỗ trợ phân trang."
      tags: [Messages]
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: page
          in: query
          schema:
            type: integer
            default: 0
          description: "Số trang (bắt đầu từ 0)."
        - name: size
          in: query
          schema:
            type: integer
            default: 50
          description: "Kích thước trang."
        - name: sort
          in: query
          schema:
            type: string
            default: "DESC"
            enum: [ASC, DESC]
          description: "Thứ tự sắp xếp theo thời gian."
      responses:
        '200':
          description: "Messages retrieved successfully."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_MessageListResponse'
              example:
                success: true
                message: "Messages retrieved successfully"
                data:
                  - id: 8
                    conversationId: 1
                    senderId: 2
                    senderUsername: "user2"
                    senderFullName: "Tran Van Binh"
                    content: "Tao phải đỡ cho em tao"
                    type: "TEXT"
                    reactions: {}
                    sentAt: "2025-12-14T06:29:24.691046Z"
                    createdAt: "2025-12-14T06:29:24.691048Z"
                    updatedAt: "2025-12-14T06:29:24.691049Z"
                    edited: false
                    deleted: false
                    forwarded: false
                    forwardCount: 0
                    readCount: 0
                  - id: 5
                    conversationId: 1
                    senderId: 2
                    senderUsername: "user2"
                    senderFullName: "Tran Van Binh"
                    content: "Hì hì đúng thật, cũng đã nhiều năm trôi qua"
                    type: "TEXT"
                    reactions: {}
                    sentAt: "2025-12-14T06:18:25.432835Z"
                    createdAt: "2025-12-14T06:18:25.432836Z"
                    updatedAt: "2025-12-14T06:18:25.432837Z"
                    edited: false
                    deleted: false
                    forwarded: false
                    forwardCount: 0
                    readCount: 1
                    readBy:
                      - userId: 1
                        username: "user1"
                        fullName: "Nguyen Linh La"
                        avatarUrl: "https://res.cloudinary.com/dk3gud5kq/image/upload/v1765476687/avatars/1.png"
                        readAt: "2025-12-14T06:50:24.557Z"
        '401':
          description: "Unauthorized - Invalid or expired token."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or expired token"
                requestId: "req-msgs001"
        '403':
          description: "Forbidden - Not a member of this conversation."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "FORBIDDEN"
                  message: "You are not a member of this conversation"
                requestId: "req-msgs002"
        '404':
          description: "Conversation not found."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "RESOURCE_NOT_FOUND"
                  message: "Conversation not found"
                requestId: "req-msgs003"
    post:
      summary: "Send a message"
      description: "Gửi một tin nhắn mới vào cuộc trò chuyện."
      tags: [Messages]
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageRequest'
      responses:
        '201':
          description: "Message sent successfully."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_MessageResponse'
              example:
                success: true
                message: "Message sent successfully"
                data:
                  id: 25
                  conversationId: 1
                  senderId: 1
                  senderUsername: "user1"
                  senderFullName: "Nguyen Linh La"
                  content: "Hello, how are you?"
                  type: "TEXT"
                  reactions: {}
                  sentAt: "2025-12-14T06:53:17.372188410Z"
                  createdAt: "2025-12-14T06:53:17.372191027Z"
                  updatedAt: "2025-12-14T06:53:17.372191325Z"
                  edited: false
                  deleted: false
                  forwarded: false
                  forwardCount: 0
                  readCount: 0
        '400':
          description: "Bad Request - Validation failed."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                empty_content:
                  summary: "Empty message content"
                  value:
                    success: false
                    error:
                      code: "VALIDATION_ERROR"
                      message: "Validation failed"
                      details:
                        - field: "content"
                          code: "NOT_BLANK"
                          message: "Content is required"
                    requestId: "req-msg001"
                invalid_reply:
                  summary: "Reply to non-existent message"
                  value:
                    success: false
                    error:
                      code: "RESOURCE_NOT_FOUND"
                      message: "Reply message not found: 99999"
                    requestId: "req-msg002"
                invalid_mentions:
                  summary: "Mention non-existent user"
                  value:
                    success: false
                    error:
                      code: "VALIDATION_ERROR"
                      message: "Validation failed"
                      details:
                        - field: "mentionedUserIds"
                          code: "VALID_USER_IDS"
                          message: "Mentioned user not found: 88888"
                    requestId: "req-msg003"
        '401':
          description: "Unauthorized - Invalid or expired token."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or expired token"
                requestId: "req-msg004"
        '403':
          description: "Forbidden - Not a member of this conversation."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "FORBIDDEN"
                  message: "You are not a member of this conversation"
                requestId: "req-msg005"
        '404':
          description: "Conversation not found."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "RESOURCE_NOT_FOUND"
                  message: "Conversation not found"
                requestId: "req-msg006"
  /v1/conversations/{conversationId}/messages/{messageId}:
    get:
      summary: "Get a specific message"
      description: "Lấy thông tin chi tiết của một tin nhắn cụ thể."
      tags: [Messages]
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: messageId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: "Message retrieved successfully."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse_MessageResponse' }
              example:
                success: true
                message: "Message retrieved successfully"
                data:
                  id: 5
                  conversationId: 1
                  senderId: 2
                  senderUsername: "user2"
                  senderFullName: "Tran Van Binh"
                  content: "Hì hì đúng thật, cũng đã nhiều năm trôi qua"
                  type: "TEXT"
                  reactions: {}
                  sentAt: "2025-12-14T06:18:25.432835Z"
                  createdAt: "2025-12-14T06:18:25.432836Z"
                  updatedAt: "2025-12-14T06:18:25.432837Z"
                  edited: false
                  deleted: false
                  forwarded: false
                  forwardCount: 0
                  readCount: 1
                  readBy:
                    - userId: 1
                      username: "user1"
                      fullName: "Nguyen Linh La"
                      avatarUrl: "https://res.cloudinary.com/dk3gud5kq/image/upload/v1765476687/avatars/1.png"
                      readAt: "2025-12-14T06:50:24.557Z"
        '401':
          description: "Unauthorized - Invalid or expired token."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or expired token"
                requestId: "req-getmsg001"
        '403':
          description: "Forbidden - Not a member of this conversation."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "FORBIDDEN"
                  message: "You are not a member of this conversation"
                requestId: "req-getmsg002"
        '404':
          description: "Message or Conversation not found."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                message_not_found:
                  summary: "Message not found"
                  value:
                    success: false
                    error:
                      code: "RESOURCE_NOT_FOUND"
                      message: "Message not found"
                    requestId: "req-getmsg003"
                conversation_not_found:
                  summary: "Conversation not found"
                  value:
                    success: false
                    error:
                      code: "RESOURCE_NOT_FOUND"
                      message: "Conversation not found"
                    requestId: "req-getmsg004"
    put:
      summary: "Update a message"
      description: "Cập nhật nội dung của một tin nhắn đã gửi. Chỉ người gửi mới có thể sửa tin nhắn."
      tags: [Messages]
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: messageId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateMessageRequest' }
      responses:
        '200':
          description: "Message updated successfully."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse_MessageResponse' }
              example:
                success: true
                message: "Message edited successfully"
                data:
                  id: 3
                  conversationId: 1
                  senderId: 1
                  senderUsername: "user1"
                  senderFullName: "Nguyen Linh La"
                  content: "da thay doi"
                  type: "TEXT"
                  sentAt: "2025-12-14T06:10:21.019500Z"
                  edited: true
                  editedAt: "2025-12-14T06:51:40.617582834Z"
                  deleted: false
                  forwarded: false
        '400':
          description: "Bad Request - Validation failed."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                empty_content:
                  summary: "Empty message content"
                  value:
                    success: false
                    error:
                      code: "VALIDATION_ERROR"
                      message: "Validation failed"
                      details:
                        - field: "content"
                          code: "NOT_BLANK"
                          message: "Content is required"
                    requestId: "req-editmsg001"
                message_deleted:
                  summary: "Cannot edit deleted message"
                  value:
                    success: false
                    error:
                      code: "BUSINESS_ERROR"
                      message: "Cannot edit a deleted message"
                    requestId: "req-editmsg002"
        '401':
          description: "Unauthorized - Invalid or expired token."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or expired token"
                requestId: "req-editmsg003"
        '403':
          description: "Forbidden - Not the message sender."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "FORBIDDEN"
                  message: "You can only edit your own messages"
                requestId: "req-editmsg004"
        '404':
          description: "Message not found."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "RESOURCE_NOT_FOUND"
                  message: "Message not found"
                requestId: "req-editmsg005"
    delete:
      summary: "Delete a message"
      description: "Xóa một tin nhắn đã gửi. Chỉ người gửi mới có thể xóa tin nhắn."
      tags: [Messages]
      security:
        - bearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: messageId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: "Message deleted successfully."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
              example:
                success: true
                message: "Message deleted successfully"
        '401':
          description: "Unauthorized - Invalid or expired token."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or expired token"
                requestId: "req-delmsg001"
        '403':
          description: "Forbidden - Not the message sender."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "FORBIDDEN"
                  message: "You can only delete your own messages"
                requestId: "req-delmsg002"
        '404':
          description: "Message not found."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "RESOURCE_NOT_FOUND"
                  message: "Message not found"
                requestId: "req-delmsg003"

  # --- Message Reactions ---
...
  /v1/read-receipts/messages/{messageId}:
    post:
      summary: "Mark message as read"
      description: "Đánh dấu một tin nhắn cụ thể là đã đọc. Tự động tạo read receipt."
      tags: [Messages]
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema: { type: integer, format: int64 }
          description: "ID của tin nhắn cần đánh dấu đã đọc"
      responses:
        '204':
          description: "Message marked as read successfully (no content)."
        '400':
          description: "Bad Request - Not a participant in this conversation."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "BAD_REQUEST"
                  message: "You are not a participant in this conversation"
                requestId: "req-read001"
        '401':
          description: "Unauthorized - Invalid or expired token."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or expired token"
                requestId: "req-read002"
        '404':
          description: "Message not found."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "RESOURCE_NOT_FOUND"
                  message: "Message not found"
                requestId: "req-read003"
    get:
      summary: "Get read receipts for message"
      description: "Lấy danh sách người dùng đã đọc tin nhắn này, bao gồm thời gian đọc."
      tags: [Messages]
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema: { type: integer, format: int64 }
          description: "ID của tin nhắn"
      responses:
        '200':
          description: "Read receipts retrieved successfully."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse_ReadReceiptListResponse' }
              example:
                success: true
                message: "Read receipts retrieved successfully"
                data:
                  - userId: 1
                    username: "user1"
                    fullName: "Nguyen Linh La"
                    avatarUrl: "https://res.cloudinary.com/dk3gud5kq/image/upload/v1765476687/avatars/1.png"
                    readAt: "2025-12-14T06:50:24.557Z"
                  - userId: 3
                    username: "hoale"
                    fullName: "Le Thi Hoa"
                    avatarUrl: "https://res.cloudinary.com/dk3gud5kq/image/upload/v1765476687/avatars/3.png"
                    readAt: "2025-12-14T07:15:30.123Z"
        '400':
          description: "Bad Request - Not a participant in this conversation."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "BAD_REQUEST"
                  message: "You are not a participant in this conversation"
                requestId: "req-getread001"
        '401':
          description: "Unauthorized - Invalid or expired token."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or expired token"
                requestId: "req-getread002"
        '404':
          description: "Message not found."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "RESOURCE_NOT_FOUND"
                  message: "Message not found"
                requestId: "req-getread003"
...